requirements.txt
asgiref==3.5.1
Django==4.0.4
django-filter==21.1
mysql==0.0.3
mysql-connector-python==8.0.29
mysqlclient==2.1.0
protobuf==3.20.1
python-dotenv==0.20.0
sqlparse==0.4.2
tzdata==2022.1